requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.31
werkzeug==3.0.3
pillow==10.4.0
deepface==0.0.92
tf-keras==2.16.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
groq==0.9.0
twilio==9.2.4
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
